css/selectors/media/media-loading-state-timing.html

<!DOCTYPE html>
<title>Media Loading State: the :buffering and :stalled pseudo-classes</title>
<link rel="help" href="https://drafts.csswg.org/selectors/#media-loading-state">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<body>
  <video width="300" height="300" muted autoplay></video>
  <script>
    // The order of "is currently stalled" changing and events firing changed in
    // https://github.com/whatwg/html/pull/11959. This test asserts that when
    // the stalled event is fired, :stalled already matches.
    async_test((t) => {
      assert_implements(CSS.supports("selector(:stalled)"), ":stalled is not supported");
      const video = document.querySelector("video");
      // Send the first 100 kB and then stall "forever" (1 hour).
      video.src = `/media/movie_300.webm?pipe=trickle(100000:d3600)&random=${Math.random()}`;
      video.addEventListener("stalled", t.step_func_done(() => {
        assert_true(video.matches(":stalled"), ":stalled matches in stalled event handler");
      }));
    }, "Test :stalled timing relative to stalled event");

    // Note: it would be good to also test that :stalled doesn't match when the
    // progress event is fired, but that would require resuming the loading
    // again on demand and pipe=trickle doesn't support this.
  </script>
</body>
